Dodgeball Game Uploaded
This program is a game I made a year ago and everything in it I built from scratch with the help from a couple of tutorials. I made the game loop and the hitboxs myself which took a lot more time then one would think. Simple game that adds more dodgeballs for the white square to dodge every 5 seconds. Move with arrow keys. Z to start. Best score was 1:30.