Add files via upload
The .jar file to run "Dodgeball".
Use the arrow keys to move the white square around to avoid the red balls. Can only stay within the blue boundaries.
The goal is to not be hit for as long as possible. A timer is at the top left corner to tell you how long you've been in the game.
When finally hit (and you will be hit), the game will tell you your score. My best score was 108 seconds. Good luck!